requirements.txt
google-generativeai>=0.4.0
google-cloud-aiplatform>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
